Relocated descriptor classes
