+++ interop-qa-tests/src/test/java/it/pagopa/pn/interop/cucumber/steps/m2m/eservice/descriptor/assistant/EServiceDescriptorPatchContext.java
@@ -1,4 +1,4 @@
-package it.pagopa.pn.interop.cucumber.steps.m2m.eservice.assistant.descriptor;
+package it.pagopa.pn.interop.cucumber.steps.m2m.eservice.descriptor.assistant;
 
 import io.cucumber.spring.ScenarioScope;
 import it.pagopa.interop.generated.openapi.clients.m2mGateway.model.EServiceDescriptor;
