+++ interop-qa-tests/src/test/java/it/pagopa/pn/interop/cucumber/steps/m2m/eservice/descriptor/assistant/EServiceDescriptorPatchOperationsAssistant.java
@@ -1,4 +1,4 @@
-package it.pagopa.pn.interop.cucumber.steps.m2m.eservice.assistant.descriptor;
+package it.pagopa.pn.interop.cucumber.steps.m2m.eservice.descriptor.assistant;
 
 import io.cucumber.spring.ScenarioScope;
 import it.pagopa.interop.eservice.service.IM2MEserviceDescriptorClient.EServiceDescriptorPatchRequest;
@@ -6,7 +6,7 @@ import it.pagopa.interop.generated.openapi.clients.m2mGateway.model.AgreementApp
 import it.pagopa.interop.generated.openapi.clients.m2mGateway.model.EServiceDescriptor;
 import it.pagopa.pn.interop.cucumber.steps.ClientTokenConfigurator;
 import it.pagopa.pn.interop.cucumber.steps.SharedStepsContext;
-import it.pagopa.pn.interop.cucumber.steps.m2m.eservice.mapper.descriptor.EServiceDescriptorMapper;
+import it.pagopa.pn.interop.cucumber.steps.m2m.eservice.descriptor.mapper.EServiceDescriptorMapper;
 import java.util.List;
 import java.util.UUID;
 import lombok.EqualsAndHashCode;
